Buttons to mass enable or disable scheduled backups https://virtualmin.com/node/25795

+++ delete_scheds.cgi
@@ -7,16 +7,44 @@ require './virtual-server-lib.pl';
 @d = split(/\0/, $in{'d'});
 @d || &error($text{'dsched_enone'});
 
-# Do the deletion
-@scheds = &list_scheduled_backups();
+# Get the backups to operate on
+@allscheds = &list_scheduled_backups();
 foreach $sid (@d) {
-	($sched) = grep { $_->{'id'} eq $sid } @scheds;
+	($sched) = grep { $_->{'id'} eq $sid } @allscheds;
 	$sched || &error(&text('dsched_egone', $sid));
-	&can_backup_sched($sched) || &error(&text('dsched_ecannot', $sid));
-	&delete_scheduled_backup($sched);
+	&can_backup_sched($sched) ||
+		&error(&text('dsched_ecannot', $sid));
+	push(@scheds, $sched);
 	}
 
-&run_post_actions_silently();
-&webmin_log("delete", "scheds", scalar(@d));
+if ($in{'disable'}) {
+	# Disable selected
+	foreach $sched (@scheds) {
+		if ($sched->{'enabled'}) {
+			$sched->{'enabled'} = 0;
+			&save_scheduled_backup($sched);
+			}
+		}
+	}
+elsif ($in{'enable'}) {
+	foreach $sched (@scheds) {
+		if (!$sched->{'enabled'}) {
+			$sched->{'enabled'} = 1;
+			&save_scheduled_backup($sched);
+			}
+		}
+	}
+elsif ($in{'delete'}) {
+	# Do the deletion
+	foreach $sched (@scheds) {
+		&delete_scheduled_backup($sched);
+		}
+	&run_post_actions_silently();
+	&webmin_log("delete", "scheds", scalar(@d));
+	}
+else {
+	&error("No button clicked");
+	}
+
 &redirect("list_sched.cgi");
